requirements.txt
anthropic==0.8.1
manifold3d==1.1.1
numpy==1.24.3
Pillow==10.0.0
trimesh==3.23.5
polyscope==1.3.0 